+++ libs/langchain_v1/tests/unit_tests/agents/middleware/test_middleware_diagram.py
@@ -1,0 +1,193 @@
+from collections.abc import Callable
+
+from syrupy.assertion import SnapshotAssertion
+
+from langchain.agents.factory import create_agent
+from langchain.agents.middleware.types import AgentMiddleware, ModelRequest
+from langchain_core.messages import AIMessage
+
+from ..model import FakeToolCallingModel
+
+
+def test_create_agent_diagram(
+    snapshot: SnapshotAssertion,
+):
+    class NoopOne(AgentMiddleware):
+        def before_model(self, state, runtime):
+            pass
+
+    class NoopTwo(AgentMiddleware):
+        def before_model(self, state, runtime):
+            pass
+
+    class NoopThree(AgentMiddleware):
+        def before_model(self, state, runtime):
+            pass
+
+    class NoopFour(AgentMiddleware):
+        def after_model(self, state, runtime):
+            pass
+
+    class NoopFive(AgentMiddleware):
+        def after_model(self, state, runtime):
+            pass
+
+    class NoopSix(AgentMiddleware):
+        def after_model(self, state, runtime):
+            pass
+
+    class NoopSeven(AgentMiddleware):
+        def before_model(self, state, runtime):
+            pass
+
+        def after_model(self, state, runtime):
+            pass
+
+    class NoopEight(AgentMiddleware):
+        def before_model(self, state, runtime):
+            pass
+
+        def after_model(self, state, runtime):
+            pass
+
+    class NoopNine(AgentMiddleware):
+        def before_model(self, state, runtime):
+            pass
+
+        def after_model(self, state, runtime):
+            pass
+
+    class NoopTen(AgentMiddleware):
+        def before_model(self, state, runtime):
+            pass
+
+        def wrap_model_call(
+            self,
+            request: ModelRequest,
+            handler: Callable[[ModelRequest], AIMessage],
+        ) -> AIMessage:
+            return handler(request)
+
+        def after_model(self, state, runtime):
+            pass
+
+    class NoopEleven(AgentMiddleware):
+        def before_model(self, state, runtime):
+            pass
+
+        def wrap_model_call(
+            self,
+            request: ModelRequest,
+            handler: Callable[[ModelRequest], AIMessage],
+        ) -> AIMessage:
+            return handler(request)
+
+        def after_model(self, state, runtime):
+            pass
+
+    agent_zero = create_agent(
+        model=FakeToolCallingModel(),
+        tools=[],
+        system_prompt="You are a helpful assistant.",
+    )
+
+    assert agent_zero.get_graph().draw_mermaid() == snapshot
+
+    agent_one = create_agent(
+        model=FakeToolCallingModel(),
+        tools=[],
+        system_prompt="You are a helpful assistant.",
+        middleware=[NoopOne()],
+    )
+
+    assert agent_one.get_graph().draw_mermaid() == snapshot
+
+    agent_two = create_agent(
+        model=FakeToolCallingModel(),
+        tools=[],
+        system_prompt="You are a helpful assistant.",
+        middleware=[NoopOne(), NoopTwo()],
+    )
+
+    assert agent_two.get_graph().draw_mermaid() == snapshot
+
+    agent_three = create_agent(
+        model=FakeToolCallingModel(),
+        tools=[],
+        system_prompt="You are a helpful assistant.",
+        middleware=[NoopOne(), NoopTwo(), NoopThree()],
+    )
+
+    assert agent_three.get_graph().draw_mermaid() == snapshot
+
+    agent_four = create_agent(
+        model=FakeToolCallingModel(),
+        tools=[],
+        system_prompt="You are a helpful assistant.",
+        middleware=[NoopFour()],
+    )
+
+    assert agent_four.get_graph().draw_mermaid() == snapshot
+
+    agent_five = create_agent(
+        model=FakeToolCallingModel(),
+        tools=[],
+        system_prompt="You are a helpful assistant.",
+        middleware=[NoopFour(), NoopFive()],
+    )
+
+    assert agent_five.get_graph().draw_mermaid() == snapshot
+
+    agent_six = create_agent(
+        model=FakeToolCallingModel(),
+        tools=[],
+        system_prompt="You are a helpful assistant.",
+        middleware=[NoopFour(), NoopFive(), NoopSix()],
+    )
+
+    assert agent_six.get_graph().draw_mermaid() == snapshot
+
+    agent_seven = create_agent(
+        model=FakeToolCallingModel(),
+        tools=[],
+        system_prompt="You are a helpful assistant.",
+        middleware=[NoopSeven()],
+    )
+
+    assert agent_seven.get_graph().draw_mermaid() == snapshot
+
+    agent_eight = create_agent(
+        model=FakeToolCallingModel(),
+        tools=[],
+        system_prompt="You are a helpful assistant.",
+        middleware=[NoopSeven(), NoopEight()],
+    )
+
+    assert agent_eight.get_graph().draw_mermaid() == snapshot
+
+    agent_nine = create_agent(
+        model=FakeToolCallingModel(),
+        tools=[],
+        system_prompt="You are a helpful assistant.",
+        middleware=[NoopSeven(), NoopEight(), NoopNine()],
+    )
+
+    assert agent_nine.get_graph().draw_mermaid() == snapshot
+
+    agent_ten = create_agent(
+        model=FakeToolCallingModel(),
+        tools=[],
+        system_prompt="You are a helpful assistant.",
+        middleware=[NoopTen()],
+    )
+
+    assert agent_ten.get_graph().draw_mermaid() == snapshot
+
+    agent_eleven = create_agent(
+        model=FakeToolCallingModel(),
+        tools=[],
+        system_prompt="You are a helpful assistant.",
+        middleware=[NoopTen(), NoopEleven()],
+    )
+
+    assert agent_eleven.get_graph().draw_mermaid() == snapshot
