+++ libs/langchain_v1/tests/unit_tests/agents/middleware/test_shell_tool_middleware.py
